moved lattice reader to input_readers
